+++ semana7/LabeUser/labeuser/src/components/Home.js
@@ -12,74 +12,56 @@ const headers = {
 
 class Home extends React.Component {
   state = {
-    Lnomes: [],
-    Lemail: [],
-    InputNome: "",
-    InputEmail: "",
+    inputNome: "",
+    inputEmail: "",
   };
 
-  componentDidMount = () => {
-    this.pegarUsuarios();
-  };
 
   mudaInputNome = (e) => {
-    this.setState({ InputNome: e.target.value });
+    this.setState({ inputNome: e.target.value });
   };
 
   mudaInputEmail = (e) => {
-    this.setState({ InputEmail: e.target.value });
+    this.setState({ inputEmail: e.target.value });
   };
 
-  pegarUsuarios = () => {
-    axios
-      .get(url, headers)
-      .then((res) => {
-        this.setState({ Lnomes: res.data });
-        console.log(res);
-      })
-      .catch((error) => {
-        alert(error.response.data);
-      });
-  };
 
   novoUsuario = () => {
     const body = {
-      name: this.state.InputNome,
-      email: this.state.InputEmail,
+      name: this.state.inputNome,
+      email: this.state.inputEmail,
     };
     axios
       .post(url, body, headers)
       .then((res) => {
         alert("Usuario cadastrado com sucesso!");
-        this.setState({ InputNome: "" });
-        this.setState({ InputEmail: "" });
-        this.pegarUsuarios();
+        this.setState({ inputNome: "" });
+        this.setState({ inputEmail: "" });
       })
       .catch((error) => {
         alert(error.response.data.message);
       });
   }
 
-    render() {
-      const componenteCadastro = this.state.Lnomes.map((n) => {
-        return <li key={n.id}>{n.name}</li>;
-      })
+  render() {
 
-      return (
-        <div>
-          <h1>LabeUsers</h1>
-          <input
-            value={this.state.InputNome}
-            onChange={this.mudaInputNome}
-          />
-                    <input
-            value={this.state.InputEmail}
-            onChange={this.mudaInputEmail}
-          />
-          <button onClick={this.novoUsuario}>Enviar</button>
-        </div>
-      )
-    }
+
+    return (
+      <div>
+        <h1>LabeUsers</h1>
+        <input
+          value={this.state.inputNome}
+          onChange={this.mudaInputNome}
+        />
+        <input
+          value={this.state.inputEmail}
+          onChange={this.mudaInputEmail}
+        />
+        <button onClick={this.novoUsuario}>Enviar</button>
+        <button onClick={this.props.irParaLista}>Ir para lista</button>
+      </div>
+    )
+  }
 }
 
 export default Home;
